Distinguish local and shared test helpers
